backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg[binary]==3.2.3; python_version >= "3.13"
openpyxl==3.1.5
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.6.6
gunicorn==22.0.0
Pillow==10.4.0
